requirements.txt
gradio
pandas
scikit-learn
google-cloud-storage
